Controllers y correcion en la DB

+++ backendGestorAmbulancia/src/main/java/com/vv/code/service/AmbulanciaService.java
@@ -11,7 +11,9 @@ import org.springframework.stereotype.Service;
 import com.vv.code.mapper.AmbulanciaDTOMapper;
 import com.vv.code.model.dto.AmbulanciaDTO;
 import com.vv.code.model.entity.Ambulancia;
+import com.vv.code.model.entity.Peticion;
 import com.vv.code.repository.AmbulanciaRepository;
+import com.vv.code.repository.PeticionesRepository;
 
 @Service
 /**
@@ -21,11 +23,14 @@ import com.vv.code.repository.AmbulanciaRepository;
 public class AmbulanciaService {
 
 	private final AmbulanciaRepository ambulanciaRepository;
+	private final PeticionesRepository peticionesRepository;
 	private final AmbulanciaDTOMapper ambulanciaDTOMapper;
 
-	public AmbulanciaService(AmbulanciaRepository ambulanciaRepository, AmbulanciaDTOMapper ambulanciaDTOMapper) {
+	public AmbulanciaService(AmbulanciaRepository ambulanciaRepository, PeticionesRepository peticionesRepository,
+			AmbulanciaDTOMapper ambulanciaDTOMapper) {
 		super();
 		this.ambulanciaRepository = ambulanciaRepository;
+		this.peticionesRepository = peticionesRepository;
 		this.ambulanciaDTOMapper = ambulanciaDTOMapper;
 	}
 
@@ -53,12 +58,12 @@ public class AmbulanciaService {
 		return new ResponseEntity<String>("EXITOSO", HttpStatus.CREATED);
 	}
 
-	public ResponseEntity<String> modificarAmbulancia(Long id, Boolean estado, String observaciones) {
-		if (id.equals("") || estado == null || observaciones.isEmpty()) {
+	public ResponseEntity<String> modificarAmbulancia(String id, boolean estado, String observaciones) {
+		if (id.equals("") || observaciones.isEmpty()) {
 			return new ResponseEntity<String>("FALLIDO", HttpStatus.BAD_REQUEST);
 		}
 
-		Optional<Ambulancia> ambulancia = ambulanciaRepository.findById(id);
+		Optional<Ambulancia> ambulancia = ambulanciaRepository.findById(Long.valueOf(id));
 
 		if (!ambulancia.isPresent()) {
 			return new ResponseEntity<String>("NO SE ENCUENTRA ESE RECURSO", HttpStatus.NOT_FOUND);
@@ -77,9 +82,15 @@ public class AmbulanciaService {
 			return new ResponseEntity<String>("FALLIDO", HttpStatus.BAD_REQUEST);
 		}
 
-		ambulanciaRepository.deleteById(id);
+		Optional<Ambulancia> ambulancia = ambulanciaRepository.findById(id);
 
-		return new ResponseEntity<String>("EXITOSO", HttpStatus.ACCEPTED);
+		if (ambulancia.isPresent()) {
+			ambulanciaRepository.deleteById(id);
+			return new ResponseEntity<String>("EXITOSO", HttpStatus.ACCEPTED);
+		}
+
+		return new ResponseEntity<String>("NO SE ENCUENTRA EL RECURSO", HttpStatus.NOT_FOUND);
+
 	}
 
 }
